middleware: update file names and restructure
